完善兰波顿,配置文件,玻璃等级,mixin

+++ src/main/java/me/oganesson/gregica/api/mte/IMTEChangeableBattery.java
@@ -7,9 +7,9 @@ import java.math.BigInteger;
 public interface IMTEChangeableBattery<T extends MetaTileEntity> {
     
     BigInteger getCapacity();
-    int getPassiveLoss();
+    long getPassiveLoss();
     
     BigInteger updateCapacity();
     
-    int updatePassiveLoss();
+    long updatePassiveLoss();
 }
